+++ RELEASE_NOTES.md
@@ -4,6 +4,9 @@ Release Notes
 
     * [SER-1406] - Fix create client errors
     * [SER-842] - Allow an authorised user to select an active country so that s/he can make any changes to a specific country
+    * [SER-1544]-Office dropdown in Create/Edit Client interface should be searchable
+    * [SER-1543]-Office dropdown in Import Client interface should be searchable
+    * [SER-1541]-Office dropdownlist in the transfer client interface should be searchable
 
 ## Version 1.0.0 - for use with Fineract Web App